+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileUserRepository.java
@@ -3,20 +3,40 @@ package com.sprint.mission.discodeit.repository.file;
 import com.sprint.mission.discodeit.util.FileUtil;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.stereotype.Repository;
 
 import java.io.File;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+@Repository
+@ConditionalOnProperty(
+        name = "discodeit.repository.type",
+        havingValue = "file"
+)
 public class FileUserRepository implements UserRepository {
 
-    private static final Path directoryPath = Path.of(FileUtil.getBasePath() +"/users");
+    private final Path directoryPath;
+
+    public FileUserRepository(
+            @Value("${discodeit.repository.file-directory:.discodeit}")
+            String rootDir
+    ) {
+        this.directoryPath = Paths.get(rootDir).toAbsolutePath().resolve("users");
+    }
 
     @Override
     public Optional<User> save(User user){
+        if(user == null){
+            return Optional.empty();
+        }
+
         Path filePath = Path.of(directoryPath.toAbsolutePath() + "/" + user.getId() + FileUtil.getExtension());
         FileUtil.saveEntity(filePath, user);
 
@@ -25,6 +45,10 @@ public class FileUserRepository implements UserRepository {
 
     @Override
     public Optional<User> findById(UUID userId) {
+        if(userId == null){
+            return Optional.empty();
+        }
+
         Path filePath = Path.of(directoryPath.toAbsolutePath() + "/" + userId + FileUtil.getExtension());
         return FileUtil.loadEntity(filePath, User.class);
     }
@@ -50,8 +74,13 @@ public class FileUserRepository implements UserRepository {
     }
 
     @Override
-    public void delete(User user) {
-        Path path = Path.of(directoryPath.toAbsolutePath() + "/" + user.getId() + FileUtil.getExtension());
+    public void delete(UUID userId) {
+        if(userId == null){
+            return;
+        }
+
+        Path path = Path.of(directoryPath.toAbsolutePath() + "/" + userId + FileUtil.getExtension());
+
         path.toFile().delete();
     }
 
@@ -65,5 +94,38 @@ public class FileUserRepository implements UserRepository {
                 file.delete();
             }
         }
+    }
+
+    @Override
+    public Optional<User> findByUserName(String username) {
+        File directory = new File(directoryPath.toAbsolutePath() + "/");
+        File[] files = directory.listFiles();
+
+        if(files != null){
+            for(File file :files) {
+                Path filePath = file.toPath();
+                Optional<User> userOptional = FileUtil.loadEntity(filePath, User.class);
+
+                if (userOptional.isPresent() && username.equals(userOptional.get().getUserName()))
+                    return userOptional;
+            }
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public Optional<User> findByEmail(String email) {
+        File directory = new File(directoryPath.toAbsolutePath() + "/");
+        File[] files = directory.listFiles();
+
+        if(files != null){
+            for(File file :files){
+                Path filePath = file.toPath();
+                Optional<User> userOptional = FileUtil.loadEntity(filePath, User.class);
+                if(userOptional.isPresent() && email.equals(userOptional.get().getEmail()))
+                    return userOptional;
+            }
+        }
+        return Optional.empty();
     }
 }
